pyproject.toml
[tool.poetry]
name = "ai-devs3"
version = "0.1.0"
description = ""
authors = ["domin <domin.dober@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
requests = "^2.32.3"
pydantic = "^2.9.2"
black = "^24.10.0"
mypy = "^1.13.0"
autoflake = "^2.3.1"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
